Completed code for relative pose computation with 3 way matches. In simple terms this involves a) Triangulation b) pnp.
This code have a lot of debug things for verifying correctness. For example
for a) Triangulation. Reproject the 3d points on each of the images.

b) for pnp we try to do pnp with 3d points and itz 2d points in curr. we should get an identity matrix.

More verification in progress!