Completed the Optimal Triangulation method to enforce the epipolar constraint on the point matches. https://github.com/mpkuse/optimal-triangulation
Preliminary results suggest probably less reprojection errors on curr and curr_m when using jointUndistort
which has the optimal triangulation. I do not know exactly how much time it takes. But
in a python implementation, it takes less than 5 mili-sec for about 150 point matches.

More work need to be done to verify this. Also more work needed on PnP. Mainly because
there are large amount of false-matches in the 3way matching.