NAP working with python threading
This is slightly better than 1-threaded implementation. However issues remain
in this. IN particular, python threading package is not a true multi-core
using package. It can just run multiple threads while waiting for IO or sleep.
This means that processes bound for CPU have no luck with this.

The way out is to use python multiprocessing package with its shared Queues (which
use a semaphore) to actually run various computations in its own processes.
I have tested in in an independent test code. now its time to test the whole system
with such an implementation.