Masked search for match_expand()
Another way could be selective sample. ie.
only put points in the tree which are in the mask. A mask can be obtained
marking the points D_pts_curr on a black image. Dilate this image in a say 40x40 neighbourhood.
Only put points in the mask for KD-tree instead of currently everything.

This has now been implemented. SO have fast processing time. upto 4x speedup on expand_match().

Verified and works as expected. Try more examples.